refactor: introduce validation handler (#6695)
* refactor: factorise rule handler invocation code


* fix


* fix


* refactor: introduce validation handler


* fix


---------
